% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{summarize_clusters}
\alias{summarize_clusters}
\title{Summarize cluster sizes}
\usage{
summarize_clusters(sce)
}
\arguments{
\item{sce}{A processed `SingleCellExperiment` object that contains
`colData(sce)$cluster_id`.}
}
\value{
A `data.frame` with cluster IDs and corresponding cell counts.
}
\description{
This helper function summarizes the number of cells in each cluster after
running \code{run_pipeline()}.
}
\examples{
\dontrun{
sce_processed <- run_pipeline(sce)
summarize_clusters(sce_processed)
}
}
