% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scPipeVis_app.R
\name{run_scPipeVis_app}
\alias{run_scPipeVis_app}
\title{Launch Shiny App for scPipeVis}
\usage{
run_scPipeVis_app(sce = NULL)
}
\arguments{
\item{sce}{Optional. A \code{SingleCellExperiment} object containing
scRNA-seq data. If provided, this object is passed to the Shiny app
via the global environment (as \code{user_sce}) and used instead of
the built-in demo dataset.}
}
\value{
No meaningful return value; this function is called for its
  side-effect of launching a Shiny app in the browser.
}
\description{
A function that launches the Shiny app for \code{scPipeVis}.
The purpose of this app is to demonstrate the single-cell RNA-seq
analysis pipeline implemented in the package. The app code has been
placed in \code{./inst/shiny-scripts}.
}
\details{
If a \code{SingleCellExperiment} object is supplied via the \code{sce}
argument, it will be made available to the app (as \code{user_sce})
and used as the input dataset. If \code{sce} is \code{NULL}, the app
will instead load a built-in demo dataset.
}
\examples{
\dontrun{
# Launch app with demo dataset:
scPipeVis::run_scPipeVis_app()

# Launch app with a user-supplied SingleCellExperiment:
# scPipeVis::run_scPipeVis_app(sce = my_sce)
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
