% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_pipeline.R
\name{visualise_pipeline}
\alias{visualise_pipeline}
\title{Visualize key steps of the single-cell analysis pipeline}
\usage{
visualise_pipeline(sce, colour_by = NA)
}
\arguments{
\item{sce}{A `SingleCellExperiment` object that has already been processed
by the pipeline function `run_pipeline()` and contains all the information required to create typical visualisations
that are part of analysing scRNA-seq data.}

\item{colour_by}{Optional. A character scalar naming a column in
`colData(sce)` to use for colouring cells in the PCA/UMAP plots. Defaults to
`NA`, in which case no colouring is applied.
Unsupervised clustering step is not yet implemented but extremely useful parameter after implementation of clustering}
}
\value{
The function does not explicitely return anything but is called for its
side-effect of producing plots.
}
\description{
This function produces quick diagnostic plots from a processed
`SingleCellExperiment` object. It first visualizes the mean–variance
relationship estimated during feature selection (i.e. the output of
`scran::modelGeneVar()` that was stored in `metadata(sce)$hvg_data`), and then
plots low-dimensional embeddings (PCA and UMAP). If `colour_by` is supplied,
the reduced-dimension plots will be coloured by the specified column in
`colData(sce)`.
}
\details{
This function assumes that:
\itemize{
  \item `run_pipeline()` function was run on a `SingleCellExperiment` object and strictly that
   the output of that function is the input to this function.
}
}
\examples{
\dontrun{
visualise_pipeline(sce)
visualise_pipeline(sce, colour_by = "cluster")
}

}
